Split out ApiInstance and factories into individual classes in factories package
