+++ koreai/src/main/java/org/finos/symphony/toolkit/koreai/spring/KoreAIBridgeFactoryImpl.java
@@ -13,8 +13,8 @@ import org.finos.symphony.toolkit.koreai.request.KoreAIRequester;
 import org.finos.symphony.toolkit.koreai.request.KoreAIRequesterImpl;
 import org.finos.symphony.toolkit.koreai.response.KoreAIResponseBuilder;
 import org.finos.symphony.toolkit.koreai.response.KoreAIResponseBuilderImpl;
-import org.finos.symphony.toolkit.spring.api.ApiInstance;
-import org.finos.symphony.toolkit.spring.api.ApiInstanceFactory;
+import org.finos.symphony.toolkit.spring.api.factories.ApiInstance;
+import org.finos.symphony.toolkit.spring.api.factories.ApiInstanceFactory;
 import org.finos.symphony.toolkit.spring.api.properties.IdentityProperties;
 import org.finos.symphony.toolkit.spring.api.properties.PodProperties;
 import org.finos.symphony.toolkit.spring.api.properties.SymphonyApiProperties;
