+++ koreai/src/main/java/org/finos/symphony/toolkit/koreai/spring/KoreAIConfig.java
@@ -9,8 +9,8 @@ import javax.net.ssl.TrustManagerFactory;
 import org.finos.symphony.toolkit.json.EntityJsonTypeResolverBuilder.VersionSpace;
 import org.finos.symphony.toolkit.json.ObjectMapperFactory;
 import org.finos.symphony.toolkit.koreai.response.KoreAIResponse;
-import org.finos.symphony.toolkit.spring.api.ApiInstanceFactory;
 import org.finos.symphony.toolkit.spring.api.SymphonyApiTrustManagersConfig;
+import org.finos.symphony.toolkit.spring.api.factories.ApiInstanceFactory;
 import org.finos.symphony.toolkit.spring.api.properties.SymphonyApiProperties;
 import org.finos.symphony.toolkit.stream.handler.ExceptionConsumer;
 import org.finos.symphony.toolkit.stream.handler.SymphonyStreamHandler;
