+++ shared-stream-spring-boot-starter/src/test/java/org/finos/symphony/toolkit/stream/handler/SymphonyStreamHandlerIT.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import javax.ws.rs.BadRequestException;
 
-import org.finos.symphony.toolkit.spring.api.ApiInstance;
+import org.finos.symphony.toolkit.spring.api.factories.ApiInstance;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
