+++ symphony-api-spring-boot-starter/src/main/java/org/finos/symphony/toolkit/spring/api/factories/HealthCheckingApiInstance.java
@@ -1,0 +1,22 @@
+package org.finos.symphony.toolkit.spring.api.factories;
+
+import org.finos.symphony.toolkit.spring.api.health.AgentHealthHelper;
+import org.springframework.boot.actuate.health.Health;
+import org.springframework.boot.actuate.health.HealthIndicator;
+
+class HealthCheckingApiInstance extends ApiInstanceDelegate implements HealthIndicator {
+
+	private AgentHealthHelper agentHealth;
+	
+	public HealthCheckingApiInstance(ApiInstance delegate, AgentHealthHelper agentHealth) {
+		super(delegate);
+		this.agentHealth = agentHealth;
+	}
+
+	@Override
+	public Health health() {
+		return agentHealth.health();
+	}
+	
+	
+}
