+++ symphony-api-spring-boot-starter/src/main/java/org/finos/symphony/toolkit/spring/api/factories/TokenManagingApiInstanceFactory.java
@@ -1,4 +1,4 @@
-package org.finos.symphony.toolkit.spring.api;
+package org.finos.symphony.toolkit.spring.api.factories;
 
 import java.util.ArrayList;
 import java.util.List;
