+++ symphony-app-spring-boot-starter/src/main/java/org/finos/symphony/toolkit/spring/app/obo/DefaultOboApiInstanceFactory.java
@@ -1,0 +1,7 @@
+package org.finos.symphony.toolkit.spring.app.obo;
+
+import org.finos.symphony.toolkit.spring.api.factories.DefaultApiInstanceFactory;
+
+public class DefaultOboApiInstanceFactory extends DefaultApiInstanceFactory {
+
+}
