+++ teamcity-symphony-integration/teamcity-symphony-integration-server/src/main/java/org/finos/symphony/toolkit/teamcity/SymphonyAdminController.java
@@ -15,10 +15,10 @@ import javax.servlet.http.HttpServletResponse;
 import org.apache.log4j.Logger;
 import org.finos.symphony.toolkit.json.EntityJsonTypeResolverBuilder.VersionSpace;
 import org.finos.symphony.toolkit.json.ObjectMapperFactory;
-import org.finos.symphony.toolkit.spring.api.ApiInstance;
-import org.finos.symphony.toolkit.spring.api.ApiInstanceFactory;
-import org.finos.symphony.toolkit.spring.api.TokenManagingApiInstanceFactory;
 import org.finos.symphony.toolkit.spring.api.builders.ApiBuilderFactory;
+import org.finos.symphony.toolkit.spring.api.factories.ApiInstance;
+import org.finos.symphony.toolkit.spring.api.factories.ApiInstanceFactory;
+import org.finos.symphony.toolkit.spring.api.factories.TokenManagingApiInstanceFactory;
 import org.finos.symphony.toolkit.spring.api.properties.IdentityProperties;
 import org.finos.symphony.toolkit.spring.api.properties.TrustStoreProperties.Type;
 import org.jetbrains.annotations.NotNull;
